Work Completed on 9-20-2018
Work Completed on 9-20-2018 - Finished up WPCS-Compliance-Type stuff in the class-book.php file. Modifed the Refresh_amazon_review function to  simply change the expiration date in the existing URL, instead of making a completely new API call (apparently that works). Began work on re-doing the 'Add Book' form - goal is to use this one form for both the 'Add a Book' and the 'Edit Book' sections, as opposed to the existing 2 forms.