Work completed on 11-2-2018
Additional work completed on 11-2-2018 - still working on getting the user role set up correctly, and allowing as minimal access to the dashboard as possible. Mind the code on line 44-45 in class-wpbooklist-general-functions.php, and modify the capability values in class-wpbooklist-utilities-accesscheck.php, line 133. Also this is helpful: https://api.wordpress.org/secret-key/1.1/salt/